% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{convert_to_model}
\alias{convert_to_model}
\title{Convert to Model}
\usage{
convert_to_model(model, to_add, dat, type)
}
\arguments{
\item{model}{original model}

\item{to_add}{string of the name of the factor to add}

\item{dat}{name of the dataframe}

\item{type}{type of model. Options are 'glmBinomial', 'glm', 'lmer', 'gam', 'gamBinomial'}
}
\description{
Takes a model and a factor to add and returns a new model with the additional factor to add. Used by the step_up function.
}
\examples{
convert_to_model()
}
\keyword{comparison}
\keyword{model}
