% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getNthElement}
\alias{getNthElement}
\title{Get Nth Element}
\usage{
getNthElement(string, n, delim = "_")
}
\arguments{
\item{string}{the string to split}

\item{n}{which element of the split string to return}

\item{delim}{character to split on, default is underscore}
}
\description{
Returns the nth element of a string split on a character
}
\examples{
getNthElement()
}
\keyword{strings}
