% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_latex.R
\name{getStars}
\alias{getStars}
\title{Get Stars}
\usage{
getStars(pValue)
}
\arguments{
\item{pValue}{the p-value}
}
\description{
Returns the correct number of stars for the p-value.
}
\examples{
getStars()
}
\keyword{latex}
