% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_latex.R
\name{makeBold}
\alias{makeBold}
\title{Make Bold}
\usage{
makeBold(x)
}
\arguments{
\item{x}{text to be made bold}
}
\description{
Makes some text bold in Latex.
}
\examples{
makeBold()
}
\keyword{latex}
