% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_to_latex.R
\name{modelToLatex}
\alias{modelToLatex}
\title{Model to Latex}
\usage{
modelToLatex(model, roundN = 3, boldSig = TRUE, usePrettyNames = FALSE,
  prettyNames = "", modelType = "lmer")
}
\arguments{
\item{model}{model}

\item{roundN}{number of digits to round to. Defaults to 3}

\item{boldSig}{make significant factors bold? Defaults to TRUE.}

\item{usePrettyNames}{Boolean. Instead of your variable names, use pretty names for each of the factors instead? Defaults to FALSE}

\item{prettyNames}{if usePrettyNames==TRUE, vector of names to use for variables}

\item{modelType}{type of model. Defaults to lmer, choices are 'lmer', 'glm', or 'gam'}
}
\description{
Takes a model and prints a pretty Latex table of the model. (Doesn't return anything.)
}
\examples{
modelToLatex()
}
\keyword{latex}
