% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{step_down}
\alias{step_down}
\title{Step Down}
\usage{
step_down(model, dependent, aic_only = FALSE, dat = "temp")
}
\arguments{
\item{model}{model to step down from}

\item{dependent}{name of dependent variable, as a string}

\item{aic_only}{Boolean, defaults to FALSE. Don't remember what this does.}

\item{dat}{name of the dataframe. Default to "temp"}
}
\description{
Takes a model to step down from. Build models with each of the predictors missing and compares to original model. I think only works for lmer??
}
\examples{
step_down()
}
\keyword{comparison}
\keyword{model}
